rcrutr-interviews/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
psycopg[binary]==3.1.17

# HTTP client
requests==2.31.0

# Milvus
pymilvus==2.3.6

# ML/Embeddings
sentence-transformers==2.3.1
torch==2.1.2

# OpenAI (for potential LLM-based evaluation)
openai==1.12.0

# Pydantic
pydantic==2.5.3

# Environment
python-dotenv==1.0.0

# Google Cloud (for Cloud Tasks)
google-cloud-tasks==2.15.0
google-cloud-storage==2.14.0
